refactor(pipeline): split engine/pipeline.rs into focused submodules (#565) (#611)
Break the 1493-line src/engine/pipeline.rs into a directory module so
every file sits under the 800-line ceiling enforced by CLAUDE.md:

- src/engine/pipeline/mod.rs            module wiring + PipelineResult +
                                        public re-exports
- src/engine/pipeline/color_state.rs    ColorSpace / BitDepth /
                                        TransferFn / IccState /
                                        ColorState / ColorTrackedImage +
                                        update_color_state (pub(super))
- src/engine/pipeline/capabilities.rs   OperationCapabilities +
                                        validate_operation_sequence
                                        (pub(super))
- src/engine/pipeline/optimize.rs       optimize_ops (resize/crop fusion)
- src/engine/pipeline/apply.rs          apply_ops_tracked + apply_ops
- src/engine/pipeline/tests.rs          shared helpers + color_state /
                                        operation_contract /
                                        optimize_ops test submodules
- src/engine/pipeline/tests/apply_ops.rs
                                        apply_ops test submodule
                                        (kept separate so tests.rs stays
                                        under the ceiling)

No behaviour change: the four external consumers (engine.rs,
tasks/processing.rs, stress.rs, engine.rs re-exports) keep using the
same paths (`crate::engine::pipeline::{apply_ops, apply_ops_tracked,
optimize_ops, ColorState, IccState}`). The previously-public but
currently-unused color types (BitDepth, ColorSpace, TransferFn,
ColorTrackedImage) remain re-exported under `#[allow(unused_imports)]`
so downstream Rust callers and tests can still address them by their
historical paths.

Verification:
- cargo +1.88.0 fmt --check
- cargo +1.88.0 clippy --no-default-features -- -D warnings
- cargo test --no-default-features --no-fail-fast (282 unit + 46 + 64 +
  43 + 8 property tests pass; 54 pipeline-specific tests under their new
  module paths all pass)
- npm run build && npm test (green)
